Moved content key generation to focus.ts
